+++ continual/conv.py
@@ -227,8 +227,8 @@ class _ConvCoNd(CoModule, _ConvNd):
         return output
 
     @property
-    def delay(self):
-        return self.kernel_size[0] - 1 - self.padding[0]
+    def receptive_field(self) -> int:
+        return self.kernel_size[0] + (self.kernel_size[0] - 1) * (self.dilation[0] - 1)
 
 
 class Conv1d(_ConvCoNd):